Create FillFileNamesFromList
Takes a list of file names without extensions, verifies they exist, then pastes the file name with extension in the result column.

Note: If there are file with the same name but different extension, the last extension listed in the array is passed back to the result column. In the code above, PDF will over-right (trump) tif if both File.tif and File.pdf exist.